backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
python-multipart==0.0.26
ffmpeg-python==0.2.0
openai==2.31.0
pinecone==6.0.0
python-dotenv==1.2.2
